resolve conflict -1
